+++ NEWS.md
@@ -1,3 +1,11 @@
+# folda 0.2.0
+
+* Update examples in the README and Vignette.
+* Fallback to base::svd for users that do have the latest C++ Eigen module.
+* Remove redundant help manual.
+* Export the functions needed for the `LDATree` package.
+* Update the algorithm: Introduce the Chi-Squared variable selection when no significant variables.
+
 # folda 0.1.0
 
 * Initial CRAN submission.